feat: 수정 삭제 수정

+++ src/main/java/com/avg/lawsuitmanagement/advice/service/AdviceService.java
@@ -6,11 +6,15 @@ import com.avg.lawsuitmanagement.advice.dto.AdviceDto;
 import com.avg.lawsuitmanagement.advice.repository.AdviceMapperRepository;
 import com.avg.lawsuitmanagement.advice.repository.param.*;
 import com.avg.lawsuitmanagement.common.custom.CustomRuntimeException;
+import com.avg.lawsuitmanagement.common.util.SecurityUtil;
 import com.avg.lawsuitmanagement.lawsuit.dto.LawsuitDto;
 import com.avg.lawsuitmanagement.lawsuit.repository.LawsuitMapperRepository;
+import com.avg.lawsuitmanagement.member.dto.MemberDto;
+import com.avg.lawsuitmanagement.member.repository.MemberMapperRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,14 +28,12 @@ public class AdviceService {
 
     private final AdviceMapperRepository adviceMapperRepository;
     private final LawsuitMapperRepository lawsuitMapperRepository;
-
+    private final MemberMapperRepository memberMapperRepository;
     public AdviceDto getAdviceById(long adviceId) {
         AdviceDto adviceDto = adviceMapperRepository.selectAdviceById(adviceId);
-
         if(adviceDto == null) {
             throw new CustomRuntimeException(ADVICE_NOT_FOUND);
         }
-
         return adviceDto;
     }
 
@@ -48,7 +50,6 @@ public class AdviceService {
 
         // Advice 정보를 삽입
         adviceMapperRepository.insertAdvice(InsertAdviceParam.of(form));
-
         // Advice ID를 가져옴
         Long adviceId = adviceMapperRepository.getLastInsertedAdviceId();
 
@@ -64,8 +65,6 @@ public class AdviceService {
         }
         InsertAdviceClientIdParam clientIdParam = InsertAdviceClientIdParam.of(adviceId, insertClientIdList);
         adviceMapperRepository.insertAdviceClientMap(clientIdParam);
-
-
 
         // advice_member_map 테이블에 데이터 삽입
        List<Long> formMemberIdList = form.getMemberIdList();
@@ -85,19 +84,55 @@ public class AdviceService {
 
     public void updateAdviceInfo(long adviceId, UpdateAdviceInfoForm form){
         AdviceDto adviceDto = adviceMapperRepository.selectAdviceById(adviceId);
-
         if(adviceDto == null){
             throw new CustomRuntimeException(ADVICE_NOT_FOUND);
         }
+
+        List<Long> clientIdList = lawsuitMapperRepository.selectClientByLawsuitId(form.getLawsuitId());
+        List<Long> memberIdList = lawsuitMapperRepository.selectMemberByLawsuitId(form.getLawsuitId());
+
+        List<Long> formClientIdList = form.getClientIdList();
+        List<Long> insertClientIdList = new ArrayList<>();
+
+        for(Long clientId : formClientIdList){
+            if(!clientIdList.contains(clientId)){
+                throw new CustomRuntimeException(CLIENT_NOT_FOUND_IN_LAWSUIT);
+            }
+            insertClientIdList.add(clientId);
+        }
+        InsertAdviceClientIdParam clientIdParam = InsertAdviceClientIdParam.of(adviceId, insertClientIdList);
+        adviceMapperRepository.insertAdviceClientMap(clientIdParam);
+
+        List<Long> formMemberIdList = form.getMemberIdList();
+        List<Long> insertMemberIdList = new ArrayList<>();
+        for(Long memberId : formMemberIdList){
+            if(!memberIdList.contains(memberId)){
+                throw new CustomRuntimeException(MEMBER_NOT_ASSIGNED_TO_LAWSUIT);
+            }
+            insertMemberIdList.add(memberId);
+        }
+        InsertAdviceMemberIdParam memberIdParam = InsertAdviceMemberIdParam.of(adviceId, insertMemberIdList);
+        adviceMapperRepository.insertAdviceMemberMap(memberIdParam);
         adviceMapperRepository.updateAdviceInfo(UpdateAdviceInfoParam.of(adviceId, form));
     }
 
+    @Transactional
     public void deleteAdviceInfo(long adviceId){
         AdviceDto adviceDto = adviceMapperRepository.selectAdviceById(adviceId);
 
         if(adviceDto == null){
             throw new CustomRuntimeException(ADVICE_NOT_FOUND);
         }
+        String email = SecurityUtil.getCurrentLoginEmail();
+        MemberDto loginMemberInfo = memberMapperRepository.selectMemberByEmail(email);
+
+        List<Long> memberIdList = adviceMapperRepository.selectMemberByAdviceId(adviceId);
+        if (!memberIdList.contains((loginMemberInfo.getId()))){
+            throw new CustomRuntimeException(MEMBER_NOT_ASSIGNED_TO_ADVICE);
+        }
+
         adviceMapperRepository.deleteAdviceInfo(adviceId);
+        adviceMapperRepository.deleteAdviceClientMap(adviceId);
+        adviceMapperRepository.deleteAdviceMemberMap(adviceId);
     }
 }
